requirements.txt
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.36.0
accelerate>=0.25.0
pillow>=10.0.0
numpy>=1.24.0
rembg==2.0.66  # Latest stable version
python-dotenv>=1.0.0  # For environment variables
opencv-python>=4.8.0  # Required by rembg
pymatting>=1.1.8  # Required by rembg
scikit-image>=0.21.0  # Required by rembg
onnxruntime>=1.16.3  # Required by rembg 